+++ vehicle/src/Vehicle/Compile/Sugar/Desugar.hs
@@ -410,6 +410,7 @@ elabExpr expr = case expr of
   B.Sub e1 tk e2 -> standardLibFunction "subTC" tk [e1, e2]
   B.Mul e1 tk e2 -> standardLibFunction "mulTC" tk [e1, e2]
   B.Div e1 tk e2 -> standardLibFunction "divTC" tk [e1, e2]
+  B.Pow e1 tk e2 -> builtinFunction (V.Pow V.PowRatTensor) tk [e1, e2]
   B.Min tk -> builtinFunction (V.Min V.MinRatTensor) tk []
   B.Max tk -> builtinFunction (V.Max V.MaxRatTensor) tk []
   B.Neg tk e -> builtinTypeClassOp V.NegTC tk [e]