+++ vehicle/src/Vehicle/Data/Builtin/Standard/Core.hs
@@ -243,7 +243,7 @@ instance BuiltinHasRatLiterals Builtin where
   accessDivRatTensorBuiltin = functionAccessor $ Div DivRatTensor
   accessMinRatTensorBuiltin = functionAccessor $ Min MinRatTensor
   accessMaxRatTensorBuiltin = functionAccessor $ Max MaxRatTensor
-  accessPowRatTensorBuiltin = functionAccessor PowRat
+  accessPowRatTensorBuiltin = functionAccessor $ Pow PowRatTensor
   accessReduceAddRatBuiltin = functionAccessor ReduceAddRatTensor
   accessReduceMulRatBuiltin = functionAccessor ReduceMulRatTensor
   accessReduceMinRatBuiltin = functionAccessor ReduceMinRatTensor